+++ CHANGELOG.md
@@ -13,7 +13,7 @@
     - Update actors diagram. (#148)
     - Rename UMLs.md file to standard README.md. (#150)
 - Add `Store` contract compatible with solidity `^0.8.0`. (#138)
-- Add Boost module. (#149 #151)
+- Add Boost module. (#149 #151 #153)
 
 ## v5.4.2
 - Use latest Nethermind base image `nethermindeth/nethermind:iexec`